fix: Allow for more precise widget rendering Previously, the rendering of rectangles and quads in the Sodium GUI was defined using integers. This restricted the position of widgets to each in-game pixel, meaning using a scaled GUI scale would cause the rendered position to be incorrect in some scenarios. For example, a user on a HiDPI display using a GUI scale of 8 would not see the brightness slider thumb move when setting it from 0 to 1, because it must be aligned to every in-game pixel (in this scenario, each in-game pixel is 8 real pixels on the display, because of a GUI scale of 8). To solve this, the rendering of rectangles and quads now uses doubles to define position, allowing for greater positioning precision when needed.

+++ src/main/java/me/jellysquid/mods/sodium/client/gui/widgets/AbstractWidget.java
@@ -28,7 +28,7 @@ public abstract class AbstractWidget implements IRenderable, IGuiEventListener {
         this.font.drawString(matrixStack, str, x, y, color);
     }
 
-    protected void drawRect(int x1, int y1, int x2, int y2, int color) {
+    protected void drawRect(double x1, double y1, double x2, double y2, int color) {
         float a = (float) (color >> 24 & 255) / 255.0F;
         float r = (float) (color >> 16 & 255) / 255.0F;
         float g = (float) (color >> 8 & 255) / 255.0F;
@@ -54,7 +54,7 @@ public abstract class AbstractWidget implements IRenderable, IGuiEventListener {
         RenderSystem.disableBlend();
     }
 
-    protected static void addQuad(IVertexBuilder consumer, int x1, int y1, int x2, int y2, float a, float r, float g, float b) {
+    protected static void addQuad(IVertexBuilder consumer, double x1, double y1, double x2, double y2, float a, float r, float g, float b) {
         consumer.pos(x2, y1, 0.0D).color(r, g, b, a).endVertex();
         consumer.pos(x1, y1, 0.0D).color(r, g, b, a).endVertex();
         consumer.pos(x1, y2, 0.0D).color(r, g, b, a).endVertex();
